Moved challenges to challenges folder
