+++ lib/challenges/challenge.py
