+++ lib/challenges/challenge2.py
